fix build error (lowkey thought this would be fine)

+++ src/main/java/net/oxyopia/vice/data/gui/Quad.kt
@@ -6,7 +6,7 @@ data class Quad(
 	var maxX: Float,
 	var maxY: Float
 ) {
-	fun addPadding(padding: Int): Quad {
+	fun addPadding(padding: Float): Quad {
 		minX -= padding
 		minY -= padding
